+++ src/State/ScreenProcessor.php
@@ -8,7 +8,6 @@ use ApiPlatform\Metadata\Exception\InvalidArgumentException;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProcessorInterface;
 use App\Dto\ScreenInput;
-use App\Entity\Tenant\Playlist;
 use App\Entity\Tenant\PlaylistScreenRegion;
 use App\Entity\Tenant\Screen;
 use App\Repository\PlaylistRepository;
@@ -32,7 +31,7 @@ class ScreenProcessor extends AbstractProcessor
         EntityManagerInterface $entityManager,
         ProcessorInterface $persistProcessor,
         ProcessorInterface $removeProcessor,
-        ScreenProvider $provider
+        ScreenProvider $provider,
     ) {
         parent::__construct($entityManager, $persistProcessor, $removeProcessor, $provider);
     }
@@ -41,6 +40,10 @@ class ScreenProcessor extends AbstractProcessor
     {
         // FIXME Do we really have to do (something like) this to load an existing object into the entity manager?
         $screen = $this->loadPrevious(new Screen(), $context);
+
+        if (!$screen instanceof Screen) {
+            throw new InvalidArgumentException('object must be of type Screen');
+        }
 
         assert($object instanceof ScreenInput);
         empty($object->title) ?: $screen->setTitle($object->title);
@@ -56,57 +59,67 @@ class ScreenProcessor extends AbstractProcessor
             $screen->setEnableColorSchemeChange($object->enableColorSchemeChange);
         }
 
-        // Adding relations for playlist/screen/region
-        if (isset($object->regions) && isset($screen)) {
+        // Adding relations for playlist/screen/region if object has region property.
+        if (isset($object->regions)) {
+            // Ensure regions object has valid structure
+            $this->validateRegionsAndPlaylists($object->regions);
+
+            $existingPlaylistScreenRegions = $screen->getPlaylistScreenRegions();
+
             foreach ($object->regions as $regionAndPlaylists) {
-                // Relevant region
-                $region = $this->screenLayoutRegionsRepository->findOneBy(['id' => $regionAndPlaylists['regionId']]);
+                $regionId = $regionAndPlaylists['regionId'];
+
+                $region = $this->screenLayoutRegionsRepository->findOneBy(['id' => $regionId]);
 
                 if (is_null($region)) {
-                    throw new InvalidArgumentException('Unknown region resource');
+                    throw new InvalidArgumentException(sprintf('Unknown region resource (id: %s)', $regionId));
                 }
 
-                // Collection to be saved.
-                $playlistScreenRegionCollection = new ArrayCollection();
+                $existingPlaylistScreenRegionsInRegion = $existingPlaylistScreenRegions->filter(
+                    fn (PlaylistScreenRegion $psr) => $psr->getRegion()?->getId() == $regionId
+                );
 
-                // Looping through playlists connected to region
-                foreach ($regionAndPlaylists['playlists'] as $inputPlaylist) {
-                    // Checking if playlists exists
+                $inputPlaylists = $regionAndPlaylists['playlists'];
+                $inputPlaylistIds = array_map(fn (array $entry): string => $entry['id'], $inputPlaylists);
+
+                // Remove playlist screen regions that should not exist in region.
+                /** @var PlaylistScreenRegion $existingPSR */
+                foreach ($existingPlaylistScreenRegionsInRegion as $existingPSR) {
+                    if (!in_array($existingPSR->getPlaylist()?->getId(), $inputPlaylistIds)) {
+                        $screen->removePlaylistScreenRegion($existingPSR);
+                    }
+                }
+
+                // Add or update the input playlists.
+                foreach ($inputPlaylists as $inputPlaylist) {
                     $playlist = $this->playlistRepository->findOneBy(['id' => $inputPlaylist['id']]);
 
                     if (is_null($playlist)) {
-                        throw new InvalidArgumentException('Unknown playlist resource');
+                        throw new InvalidArgumentException(sprintf('Unknown playlist resource (id: %s)', $inputPlaylist['id']));
                     }
 
-                    // See if relation already exists
-                    $existingPlaylistScreenRegion = $this->playlistScreenRegionRepository->findOneBy([
-                        'screen' => $screen,
-                        'region' => $region,
+                    $existingPlaylistScreenRegionRelation = $this->playlistScreenRegionRepository->findOneBy([
                         'playlist' => $playlist,
+                        'region' => $region,
+                        'screen' => $screen,
                     ]);
 
-                    if (is_null($existingPlaylistScreenRegion)) {
-                        // If relation does not exist, create new PlaylistScreenRegion
+                    if (!is_null($existingPlaylistScreenRegionRelation)) {
+                        $existingPlaylistScreenRegionRelation->setWeight($inputPlaylist['weight'] ?? 0);
+                    } else {
                         $newPlaylistScreenRegionRelation = new PlaylistScreenRegion();
                         $newPlaylistScreenRegionRelation->setPlaylist($playlist);
                         $newPlaylistScreenRegionRelation->setRegion($region);
                         $newPlaylistScreenRegionRelation->setScreen($screen);
                         $newPlaylistScreenRegionRelation->setWeight($inputPlaylist['weight'] ?? 0);
-                        /** @psalm-suppress InvalidArgument */
-                        $playlistScreenRegionCollection->add($newPlaylistScreenRegionRelation);
-                    } else {
-                        // Update weight, add existing relation
-                        $existingPlaylistScreenRegion->setWeight($inputPlaylist['weight'] ?? 0);
-                        /** @psalm-suppress InvalidArgument */
-                        $playlistScreenRegionCollection->add($existingPlaylistScreenRegion);
+                        $screen->addPlaylistScreenRegion($newPlaylistScreenRegionRelation);
                     }
                 }
-                $region->setPlaylistScreenRegions($playlistScreenRegionCollection);
             }
         }
 
         // Maps ids of existing groups
-        if (isset($object->groups) && isset($screen)) {
+        if (isset($object->groups)) {
             $groupCollection = new ArrayCollection();
             foreach ($object->groups as $group) {
                 $groupToSave = $this->groupRepository->findOneBy(['id' => $group]);
@@ -133,5 +146,39 @@ class ScreenProcessor extends AbstractProcessor
         }
 
         return $screen;
+    }
+
+    private function validateRegionsAndPlaylists(array $regions): void
+    {
+        foreach ($regions as $region) {
+            $this->validateRegion($region);
+
+            foreach ($region['playlists'] as $playlist) {
+                $this->validatePlaylist($playlist);
+            }
+        }
+    }
+
+    private function validateRegion(array $region): void
+    {
+        if (!isset($region['regionId']) || !is_string($region['regionId'])) {
+            throw new InvalidArgumentException('All regions must specify a valid Ulid');
+        }
+
+        if (!isset($region['playlists']) || !is_array($region['playlists'])) {
+            throw new InvalidArgumentException('All regions must specify a list of playlists');
+        }
+    }
+
+    private function validatePlaylist(array $playlist): void
+    {
+        if (!isset($playlist['id']) || !is_string($playlist['id'])) {
+            throw new InvalidArgumentException('All playlists must specify a valid Ulid');
+
+        }
+
+        if (isset($playlist['weight']) && !is_integer($playlist['weight'])) {
+            throw new InvalidArgumentException('Playlists weight must be an integer');
+        }
     }
 }